+++ packages/stream_video/lib/src/webrtc/rtc_media_device/rtc_media_device_notifier.dart
@@ -1,8 +1,11 @@
 import 'package:collection/collection.dart';
+import 'package:flutter/foundation.dart';
+import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart' as rtc;
 
 import '../../../open_api/video/coordinator/api.dart';
+import '../../call/stats/tracer.dart';
 import '../../errors/video_error_composer.dart';
 import '../../platform_detector/platform_detector.dart';
 import '../../utils/extensions.dart';
@@ -39,6 +42,13 @@ class RtcMediaDeviceNotifier {
 
   Stream<List<RtcMediaDevice>> get onDeviceChange => _devicesController.stream;
   final _devicesController = BehaviorSubject<List<RtcMediaDevice>>();
+
+  final _tracer = Tracer(null);
+
+  @internal
+  TraceSlice getTrace() {
+    return _tracer.take();
+  }
 
   /// Allows to handle call interruption callbacks.
   /// [onInterruptionStart] is called when the call interruption begins.
@@ -143,6 +153,11 @@ class RtcMediaDeviceNotifier {
           ),
       ];
 
+      _tracer.trace(
+        'navigator.mediaDevices.enumeratedevices',
+        mediaDevices.map((device) => device.toJson()).toList(),
+      );
+
       _devicesController.add(mediaDevices);
 
       if (kind != null) {
@@ -180,12 +195,14 @@ class RtcMediaDeviceNotifier {
   /// This does not affect the microphone or remote track subscriptions.
   /// Use as a global "mute all sounds" toggle or when the app goes to background.
   Future<void> pauseAudioPlayout() {
+    _tracer.trace('navigator.mediaDevices.pauseAudioPlayout', null);
     return rtc.Helper.pauseAudioPlayout();
   }
 
   /// Resumes audio output (playout) muted via [pauseAudioPlayout].
   /// Does not change microphone state or remote track subscriptions.
   Future<void> resumeAudioPlayout() {
+    _tracer.trace('navigator.mediaDevices.resumeAudioPlayout', null);
     return rtc.Helper.resumeAudioPlayout();
   }
 
@@ -195,6 +212,7 @@ class RtcMediaDeviceNotifier {
   /// To ensure you receive `onInterruptionEnd`, explicitly call
   /// [resumeAudioPlayout] (e.g., when the app resumes from background).
   Future<void> regainAndroidAudioFocus() {
+    _tracer.trace('navigator.mediaDevices.regainAndroidAudioFocus', null);
     return rtc.Helper.regainAndroidAudioFocus();
   }
 }