+++ packages/stream_video/CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fixed incoming call timeout handling.
   - Use `streamVideo.observeCoreRingingEventsForBackground()` instead of `streamVideo.observeCallDeclinedRingingEvent()` in `firebaseMessagingBackgroundHandler` to support all necessary events.
 
+* Improved SFU stats implementation.
+
 ## 1.0.1
 
 ### ✅ Added